COMMON: Uplift powerpacker as it's used by Amiga ITE as well
